Move preview component to subfolder
[MAILPOET-2811]
